chore: move cli related function out of the core
