+++ packages/cli/src/blue-harvest/domains.ts
@@ -1,10 +1,10 @@
-import { env } from '../../env';
-
 const DEFAULT_DOMAIN = 'https://app.beta.redocly.com';
 
 export function getDomain(): string {
-  if (env.REDOCLY_DOMAIN) {
-    return env.REDOCLY_DOMAIN;
+  const domain = process.env.REDOCLY_DOMAIN;
+
+  if (domain) {
+    return domain;
   }
 
   return DEFAULT_DOMAIN;
