--- packages/core/src/redocly/bh/index.ts
@@ -1,9 +1,0 @@
-import * as client from './api-client';
-import * as domains from './domains';
-import * as apiKeys from './api-keys';
-
-export namespace BlueHarvest {
-  export const ApiClient = client.ApiClient;
-  export const getDomain = domains.getDomain;
-  export const getApiKeys = apiKeys.getApiKeys;
-}
